refactor: move utils to custom-streams folder
